metamask controller - move middleware into seperate files
