+++ app/scripts/lib/createOriginMiddleware.js
@@ -1,0 +1,9 @@
+// append dapp origin domain to request
+module.exports = createOriginMiddleware
+
+function createOriginMiddleware({ origin }) {
+  return function originMiddleware (req, res, next, end) {
+    req.origin = originDomain
+    next()
+  }
+}
